+++ Moonlight/Core/Implementations/UI/Index/GreetingMessages.cs
@@ -1,11 +1,11 @@
 ﻿using MoonCoreUI.Helpers;
-using Moonlight.Core.Interfaces.UI.Index;
+using Moonlight.Core.Interfaces.UI.User;
 using Moonlight.Core.Models.Abstractions;
 using Moonlight.Core.UI.Components.Cards;
 
 namespace Moonlight.Core.Implementations.UI.Index;
 
-public class GreetingMessages : IIndexPageComponent
+public class GreetingMessages : IUserDashboardComponent
 {
     public Task<UiComponent> Get()
     {
